@@ -20,21 +20,20 @@
         }
     ],
     "require": {
-        "php": "^7.1 || ^8.0",
-        "ext-json": "*",
+        "php": "^8.1",
         "paragonie/random-lib": "^2.0",
         "symfony/polyfill-mbstring": "^1.19"
     },
     "require-dev": {
         "dg/bypass-finals": "^1.3",
-        "phpunit/phpunit": "^7.5 || ^8.5 || ^9.5"
+        "phpunit/phpunit": "^9.6"
     },
     "config": {
         "bin-dir": "bin"
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "1.0-dev"
+            "dev-master": "1.x-dev"
         }
     },
     "autoload": {
